backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tzdata==2023.3


# pip install -r requirements.txt
